+++ src/autovalaLib/autovalaLib.vala
@@ -301,6 +301,7 @@ namespace AutoVala {
 			}
 
 			globalData.generateExtraData();
+			
 			var globalElement = new ElementGlobal();
 			DataOutputStream dataStreamGlobal;
 			try {
@@ -391,6 +392,7 @@ namespace AutoVala {
 			if (error) {
 				return true;
 			}
+
 			ElementBase.globalData.clearAutomatic();
 			ElementBase.globalData.generateExtraData();
 
@@ -399,6 +401,7 @@ namespace AutoVala {
 				element.autoConfigure();
 			}
 
+			error|=ElementGResource.autoGenerate();
 			error|=ElementBashCompletion.autoGenerate();
 			error|=ElementBinary.autoGenerate();
 			error|=ElementData.autoGenerate();
@@ -821,7 +824,7 @@ namespace AutoVala {
 		 */
 		public ValaProject ? get_binaries_list(string ?basePath = null) {
 
-			this.config=new AutoVala.Configuration(basePath);
+			this.config = new AutoVala.Configuration(basePath);
 			if (config.globalData.error) {
 				return null;
 			}